miniworld datasets and dataloaders to be tested on deeplab

+++ src/pl_datamodules/isprs_vai_semisup.py
@@ -81,7 +81,7 @@ class IsprsVaiSemisup(IsprsVaiSup):
         sup_train_sampler = RandomSampler(
             data_source=self.sup_train_set,
             replacement=True,
-            num_samples=self.nb_pass_per_epoch * len(self.sup_train_set),
+            num_samples=int(self.nb_pass_per_epoch * len(self.sup_train_set)),
         )
 
         # num_workers should be the number of cpus on the machine.
@@ -97,7 +97,7 @@ class IsprsVaiSemisup(IsprsVaiSup):
         unsup_train_sampler = RandomSampler(
             data_source=self.unsup_train_set,
             replacement=True,
-            num_samples=self.nb_pass_per_epoch * len(self.unsup_train_set),
+            num_samples=int(self.nb_pass_per_epoch * len(self.unsup_train_set)),
         )
         # num_workers should be the number of cpus on the machine.
         unsup_train_dataloader = DataLoader(
